v1.4.2.1 to v1.4.2.2 (tanju)

+++ ATOOLS-2.0/Phys/HepEvt_Interface.C
@@ -1078,7 +1078,7 @@ bool HepEvt_Interface::IdentifyBlobs(ATOOLS::Blob_List * const blobs)
       }
       else {
 	ATOOLS::kf::code in=incoming->Flav().Kfcode();
-	for (size_t i=0;i<(size_t)beam->NOutP();++i) {
+	for (int i=0;i<beam->NOutP();++i) {
 	  ATOOLS::kf::code out=beam->OutParticle(i)->Flav().Kfcode();
 	  if (in==kf::e && out==kf::photon) {
 	    beam->SetType(btp::Bunch);
